+++ HISTORY.rst
@@ -4,6 +4,10 @@ History
 0.1.7 (unreleased)
 ------------------
 
+- Timestamps of all timeseries fields are shown for aggregation results.
+
+- Timestamps in the aggregation results are filtered when retrieving subsets of timeseries.
+
 - Introducing the ModelMeta class. Its main purpose at this moment is to compute all 
   possible combinations of composite_fields and aggregation variables.